# ensure we use identical ordering as the schema when we query