--- a/backend/clickhouse/logs.go
+++ b/backend/clickhouse/logs.go
@@ -78,8 +78,8 @@ func (client *Client) BatchWriteLogRows(ctx context.Context, logRows []*LogRow) 
 const LogsLimit int = 100
 const KeyValuesLimit int = 50
 
-const OrderBackward = "Timestamp ASC, UUID ASC"
-const OrderForward = "Timestamp DESC, UUID DESC"
+const OrderBackward = "toUnixTimestamp(Timestamp) ASC, UUID ASC"
+const OrderForward = "toUnixTimestamp(Timestamp) DESC, UUID DESC"
 
 type Pagination struct {
 	After     *string